+++ src/wrap/index.ts
@@ -6,29 +6,7 @@ export {
 
 export { default as transformSchema, wrapSchema } from './transformSchema';
 
-export { default as CheckResultAndHandleErrors } from './CheckResultAndHandleErrors';
-export { default as ExpandAbstractTypes } from './ExpandAbstractTypes';
-export { default as AddReplacementSelectionSets } from './AddReplacementSelectionSets';
-export { default as AddMergedTypeSelectionSets } from './AddMergedTypeSelectionSets';
-export { default as FilterToSchema } from './FilterToSchema';
-export { default as AddTypenameToAbstract } from './AddTypenameToAbstract';
-
-export { default as RenameTypes } from './RenameTypes';
-export { default as RenameRootTypes } from './RenameRootTypes';
-export { default as FilterTypes } from './FilterTypes';
-export { default as TransformRootFields } from './TransformRootFields';
-export { default as RenameRootFields } from './RenameRootFields';
-export { default as FilterRootFields } from './FilterRootFields';
-export { default as TransformObjectFields } from './TransformObjectFields';
-export { default as RenameObjectFields } from './RenameObjectFields';
-export { default as FilterObjectFields } from './FilterObjectFields';
-export { default as TransformQuery } from './TransformQuery';
-
-export { default as ExtendSchema } from './ExtendSchema';
-export { default as WrapType } from './WrapType';
-export { default as WrapFields } from './WrapFields';
-export { default as HoistField } from './HoistField';
-export { default as MapFields } from './MapFields';
+export * from './transforms/index';
 
 export {
   default as makeRemoteExecutableSchema,
@@ -36,11 +14,11 @@ export {
 } from './makeRemoteExecutableSchema';
 
 // implemented directly within initial query proxy creation
-export { default as AddArgumentsAsVariables } from './AddArgumentsAsVariables';
+export { default as AddArgumentsAsVariables } from './transforms/AddArgumentsAsVariables';
 // superseded by AddReplacementFragments
-export { default as ReplaceFieldWithFragment } from './ReplaceFieldWithFragment';
+export { default as ReplaceFieldWithFragment } from './transforms/ReplaceFieldWithFragment';
 // superseded by AddReplacementSelectionSets
-export { default as AddReplacementFragments } from './AddReplacementFragments';
+export { default as AddReplacementFragments } from './transforms/AddReplacementFragments';
 // superseded by TransformQuery
-export { default as WrapQuery } from './WrapQuery';
-export { default as ExtractField } from './ExtractField';
+export { default as WrapQuery } from './transforms/WrapQuery';
+export { default as ExtractField } from './transforms/ExtractField';
