+++ src/wrap/transforms/AddTypenameToAbstract.ts
@@ -1,7 +1,7 @@
 import { GraphQLSchema } from 'graphql';
 
-import { Transform, Request } from '../Interfaces';
-import { addTypenameToAbstract } from '../delegate/addTypenameToAbstract';
+import { Transform, Request } from '../../Interfaces';
+import { addTypenameToAbstract } from '../../delegate/addTypenameToAbstract';
 
 export default class AddTypenameToAbstract implements Transform {
   private readonly targetSchema: GraphQLSchema;
