+++ src/wrap/transforms/CheckResultAndHandleErrors.ts
@@ -1,11 +1,11 @@
 import { GraphQLSchema, GraphQLOutputType } from 'graphql';
 
-import { checkResultAndHandleErrors } from '../delegate/checkResultAndHandleErrors';
+import { checkResultAndHandleErrors } from '../../delegate/checkResultAndHandleErrors';
 import {
   Transform,
   SubschemaConfig,
   IGraphQLToolsResolveInfo,
-} from '../Interfaces';
+} from '../../Interfaces';
 
 export default class CheckResultAndHandleErrors implements Transform {
   private readonly context?: Record<string, any>;
