+++ src/wrap/transforms/ExpandAbstractTypes.ts
@@ -14,8 +14,8 @@ import {
   visitWithTypeInfo,
 } from 'graphql';
 
-import implementsAbstractType from '../utils/implementsAbstractType';
-import { Transform, Request } from '../Interfaces';
+import implementsAbstractType from '../../utils/implementsAbstractType';
+import { Transform, Request } from '../../Interfaces';
 
 type TypeMapping = { [key: string]: Array<string> };
