+++ src/wrap/transforms/TransformRootFields.ts
@@ -1,6 +1,6 @@
 import { GraphQLSchema, GraphQLField, GraphQLFieldConfig } from 'graphql';
 
-import { Transform, Request } from '../Interfaces';
+import { Transform, Request } from '../../Interfaces';
 
 import TransformObjectFields, {
   FieldNodeTransformer,
