+++ src/wrap/transforms/index.ts
@@ -1,0 +1,33 @@
+export { default as CheckResultAndHandleErrors } from './CheckResultAndHandleErrors';
+export { default as ExpandAbstractTypes } from './ExpandAbstractTypes';
+export { default as AddReplacementSelectionSets } from './AddReplacementSelectionSets';
+export { default as AddMergedTypeSelectionSets } from './AddMergedTypeSelectionSets';
+export { default as FilterToSchema } from './FilterToSchema';
+export { default as AddTypenameToAbstract } from './AddTypenameToAbstract';
+
+export { default as RenameTypes } from './RenameTypes';
+export { default as RenameRootTypes } from './RenameRootTypes';
+export { default as FilterTypes } from './FilterTypes';
+export { default as TransformRootFields } from './TransformRootFields';
+export { default as RenameRootFields } from './RenameRootFields';
+export { default as FilterRootFields } from './FilterRootFields';
+export { default as TransformObjectFields } from './TransformObjectFields';
+export { default as RenameObjectFields } from './RenameObjectFields';
+export { default as FilterObjectFields } from './FilterObjectFields';
+export { default as TransformQuery } from './TransformQuery';
+
+export { default as ExtendSchema } from './ExtendSchema';
+export { default as WrapType } from './WrapType';
+export { default as WrapFields } from './WrapFields';
+export { default as HoistField } from './HoistField';
+export { default as MapFields } from './MapFields';
+
+// implemented directly within initial query proxy creation
+export { default as AddArgumentsAsVariables } from './AddArgumentsAsVariables';
+// superseded by AddReplacementFragments
+export { default as ReplaceFieldWithFragment } from './ReplaceFieldWithFragment';
+// superseded by AddReplacementSelectionSets
+export { default as AddReplacementFragments } from './AddReplacementFragments';
+// superseded by TransformQuery
+export { default as WrapQuery } from './WrapQuery';
+export { default as ExtractField } from './ExtractField';
